face 1.5, creación del modulo de ventas y primera conección a base de datos con sqlite, creación de las vistas relacionadas con ventas e incorporación de los repositorios de sqlite

+++ src/main/java/com/perfumeria/controlador/producto/StockProductoControlador.java
@@ -91,7 +91,7 @@ public class StockProductoControlador implements ActionListener {
             if (filaSeleccionada != -1) {
                 Object valorId = tabla.getValueAt(filaSeleccionada, 0);
                 try {
-                    Long id = Long.parseLong(valorId.toString());
+                    Integer id = Integer.parseInt(valorId.toString());
                     Producto producto = Config.productoRepositorio.buscarPorId(id);
 
                     if (producto != null) {
